Finished the hard part
Now we just need to see who is able to use it or not.
i tested it, excess food is taken first (you know the food rework) and then the normal food is taken.
Known issues are the fact that if you stick later into the queue (more than one turn to build) it'll drain your food each turn until it is built, only giving one pop (same problem than with chain gang, does not seem fixable). So i think it's fine