Attempt at fixing vineship, hunter, and defender
Vineship uses "Unfallen" symbol, which isn't the same as the "root creator" icon it uses elsewhere but gets the point across.
Hunter uses "Duel" symbol and the medium-size battle button as an icon, which looks similar but is kind of the wrong resolution so isn't perfect.
Defender uses "plating" symbol and "hullplatingsmall" icon, which seems to wrok great.