Dynamic manpower from food and fix of vanilla bug.
There it goes, I did a lot of testing in a modded vanilla version, but also a fair bit in ESG, so it should be stable a working with no issues.

The bug is now fixed, this means you don't get free refills, this happened when your Empire Manpower was full and it changed briefly into full conversion, did the refills for free, as the food was already computated in systems.

Now aside the food computation, the conversion you see on the tooltip will be stored internally and used on the Empire Manpower computations. What you see is what you get next turn, and every bit of manpower will be properly taken from your manpower sources (mainly food conversion and minor civs ^^).

On the bright side we have the dynamic ratio of food sent. If you need to cover 400 manpower and you can produce 2000 from food, only 20% will be converted from food, keeping the manpower from food expenditures more fluid, aside the 5% overcorrection, that will mean in practice much, much less wasted overall.

About the code:

The meat and potatoes is on ResourceDefinitions, SimulationDescriptors[ColonizedStarSystem] and SimulationDescriptors[Empire]. All the other changes are mostly due to the need or an alternative Manpower property, EmpireManpowerBalance, to avoid any kind of circular loops, also EmpireManpowerPercent on the three effects that use percents instead of additions.

In ResourceDefinitions, we added a new stock resource, for internal purposes only: store the food conversion used when you clicked end of turn. TrueNetEmpireManpower is the new NetEmpireManpower operating behind the scenes.

In SimulationDescriptors[ColonizedStarSystem] "ShouldTakeGrowthForManpower" gets replaced with GrowthForManpowerRatio, the first was the on-off switch from vanilla, the latter will storage the percentage used for the dynamic conversion, at the same moment food is transferred. The maximum capacity of conversion is calculated here, and the ratio is applied also here.

SimulationDescriptors[Empire]: We get a bunch of new properties for internal calculations most of it: MaximumEnrollmentRate gets the feedback form colonies using the Composition="Sum" property, the storage of the conversion quantity used is done here, also the calculation of the ratio of conversion needed to cover the missing empire manpower.