Fix to flotilla shield not being applied properly.
Fixes flotilla shield so it properly adds the extra shield to all ships in the flotilla.

The Battle Action "Group_SetInitialMoraleBonus" is triggered at battle start, all we need to do, is to call "Ship_ResetShield" Battle Effect for all the ships with shield, so we update the ships with the new shield cap provided by the flotilla shield.

Fixes #1590