Manpower from food fixes.
Fixes Percent Empire Manpower Effects like Microstasis Motors being accounted for properly when calculating Manpower for food.

Adds the effects from Luxuries that modify the manpower output from systems to also be accounted for when calculating the maximum manpower from food.

Fixed an issue with the calculation of the Empire Manpower Upkeep, as modifies the percent for the calculations of manpower from food the moment is calculated, as it removes the upkeep and refreshes the simulation modifiers.